+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- [added] Add "Edit" button with editor selection, open local computer code editors.
 
 ## [v0.4.0] - 2026-03-01